+++ client/src/pages/AddBusiness/AddBusinessOld.js
@@ -12,21 +12,27 @@ class AddBusiness extends Component {
   constructor () {
     super ();
     this.state = {
-      googleID: "",
       name: "",
       day: "",
       beginTime: "",
       endTime: "",
-      info: "",
-      results: [],
-      search: ""
+      info: ""
     };
     this.handleInputChange.bind(this);
   }
 
+  // state = {
+  //   businesses: [],
+  //   name: "",
+  //   // days: [], // Not sure if this should be an array - Stef
+  //   day: "",
+  //   beginTime: "", // Not sure how to input if there are multiple start and end times (multiple happy hours in a day) - Stef
+  //   endTime: "",
+  //   info: ""
+  // };
+
   componentDidMount() {
     console.log("loaded");
-    // Run API.getPlaces by current location + search term?
   }
 
   handleInputChange = event => {
@@ -35,64 +41,22 @@ class AddBusiness extends Component {
     });
   };
 
-  handleSelectedOption = event => {
-    const select = event.target;
-    const selectedOption = select[select.selectedIndex];
-    const optionName = selectedOption.getAttribute('value').split(",")[0];
-    const optionID = selectedOption.getAttribute('value').split(",")[1];
-    this.setState({
-      name: optionName,
-      googleID: optionID
-    });
-    console.log(optionName, optionID);
-  }
-
-  handleSearchSubmit = event => {
-    event.preventDefault();
-    this.searchGoogle(this.state.search);
-  };
-
-  searchGoogle(query) {
-    console.log("google has been searched")
-    API.getPlaces(query)
-    // .then(
-    //   function (searchResults) {
-    //     console.log(searchResults.data);
-    //   })
-    .then(res =>
-        this.setState({
-          results: res.data.results
-        })
-      )
-      .catch(err => console.log(err));
-  };
-
   handleFormSubmit = event => {
     event.preventDefault();
-    if (! this.state.name) {
+    if (! this.state.name || ! this.state.day || ! this.state.info) {
       alert("All fields must be filled out");
     } else {
       API.saveBusiness({
         name: this.state.name,
-        googleID: this.state.googleID
-        // day: this.state.day,
-        // beginTime: this.state.beginTime,
-        // endTime: this.state.endTime,
-        // info: this.state.info
+        day: this.state.day,
+        beginTime: this.state.beginTime,
+        endTime: this.state.endTime,
+        info: this.state.info
       })
-        // .then(
-        //   API.saveDeal({
-        //     googleID: this.state.googleID,
-        //     day: this.state.day,
-        //     beginTime: this.state.beginTime,
-        //     endTime: this.state.endTime,
-        //     info: this.state.info
-        //   })
-        // )
         .then(
           function (newDeal) {
             console.log(newDeal.data);
-            alert("Thanks for the info, pal!");
+            alert("Thanks for the info, pal!")
           },
           this.setState({
               name: "",
@@ -109,48 +73,21 @@ class AddBusiness extends Component {
   render() {
     return (
       <div>
-      <SearchBar
-        onClick={this.handleSearchSubmit}
-        onChange={this.handleInputChange}
-      />
+      <SearchBar />
       <Container>
         <Row>
           <Col size="md-6">
             <h5 className="title">Add New Business</h5>
             <form>
 
-              {/* ***Input business name*** */}
-              {/* <label htmlFor="businessName">Enter business name.</label>
+              <label htmlFor="businessName">Enter business name.</label>
               <Input
                 value={this.state.name}
                 onChange={this.handleInputChange}
                 name="name"
                 placeholder="Business Name"
-              /> */}
+              />
 
-              {/* ***Select business name(after search)*** */}
-              <div className="form-group">
-              <label htmlFor="business">Business Name:</label>
-              {/* If businesses exist in the database: */}
-              {this.state.results.length ? (
-                <select onChange={this.handleSelectedOption} defaultValue="">
-                  <option value="" disabled>Select your option</option>
-                  {this.state.results.map(place => (
-                    <option
-                      key={place.id}
-                      value={[place.name,place.id]}
-                    >
-                      {place.name}
-                    </option>
-                  ))}
-                </select>
-              // Default message if no business exists in the database.
-              ) : (
-                <h3>Search for the name or type of the business.</h3>
-              )}
-              </div>
-
-              {/* ***Select day of deal*** */}
               <label>
                 Choose a day from this list:
                 <select
@@ -160,17 +97,16 @@ class AddBusiness extends Component {
                   selected={this.state.day}
                   onChange={this.handleInputChange}
                 >
-                  <option value="1">Sunday</option>
-                  <option value="2">Monday</option>
-                  <option value="3">Tuesday</option>
-                  <option value="4">Wednesday</option>
-                  <option value="5">Thursday</option>
-                  <option value="6">Friday</option>
-                  <option value="7">Saturday</option>
+                  <option value="Sunday">Sunday</option>
+                  <option value="Monday">Monday</option>
+                  <option value="Tuesday">Tuesday</option>
+                  <option value="Wednesday">Wednesday</option>
+                  <option value="Thursday">Thursday</option>
+                  <option value="Friday">Friday</option>
+                  <option value="Saturday">Saturday</option>
                 </select>
               </label>
 
-              {/* ***Input start & end times for the deal*** */}
               <label htmlFor="dealTimes">Select the start and end times of the deal.</label>
               <div className="form-row">
                 <Input
@@ -179,6 +115,8 @@ class AddBusiness extends Component {
                     type="time"
                     id="appt-time"
                     name="beginTime"
+                    min="9:00"
+                    max="18:00"
                     required
                 />
                 <p className="timeframe">to</p>
@@ -188,11 +126,12 @@ class AddBusiness extends Component {
                     type="time"
                     id="appt-time"
                     name="endTime"
+                    min="9:00"
+                    max="18:00"
                     required
                 />
               </div>
 
-              {/* ***Describe the deal*** */}
               <label htmlFor="description">Enter a description of the deal.</label>
               <TextArea
                 // className="description"
@@ -203,7 +142,6 @@ class AddBusiness extends Component {
                 onChange={this.handleInputChange}
               />
 
-              {/* ***Submit the new business into the database with its first deal*** */}
               <FormBtn
                 className="btn-primary"
                 onClick={this.handleFormSubmit}
