+++ server/db/models/business.js
@@ -2,11 +2,21 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 var businessSchema = new Schema({
-    name: { type: String, required: true },
-    day: { type: String, required: true },
-    beginTime: { type: String, required: true },
-    endTime: { type: String, required: true },
-    info: { type: String, required: true },
+    googleID: { type: String, unique: true },
+    name: { type: String, unique: true, required: true },
+    // address: { type: String, required: true },
+    reviews: [
+        {
+            type: Schema.Types.ObjectId,
+            ref: "Review"
+        }
+    ],
+    deals: [
+        {
+            type: Schema.Types.ObjectId,
+            ref: "Deal"
+        }
+    ],
     updated_date: { type: Date, default: Date.now },
 });
